Refactor around the location of where runtimeType.container is initialized
Specifically, this moves the files:
executor/runtime/container.go -> executor/runtime/types/container.go
executor/runtime/container_test.go -> executor/runtime/types/container_test.go

The job of initializing the Container object should be done not in a different
package as where the Container type is declarared, but the same one.
